+++ src/transformers/song-transformer.ts
@@ -16,8 +16,8 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import type { SongDTO, AlbumDTO, ArtistDTO, PlaylistDTO } from '../types/index.js';
-
+import type { SongDTO } from '../types/index.js';
+import { formatDuration, extractGenre, extractAllGenres } from './shared-transformers.js';
 
 /**
  * Raw song data from Navidrome API
@@ -42,102 +42,6 @@ export interface RawSong {
   createdAt?: string;
   path?: string;
   [key: string]: unknown; // Allow other fields we don't use
-}
-
-export interface RawAlbum {
-  id: string;
-  name: string;
-  artist: string;
-  artistId: string;
-  albumArtist?: string;
-  albumArtistId?: string;
-  releaseYear?: number;
-  genre?: string;
-  genres?: Array<{ id: string; name: string }>;
-  songCount: number;
-  duration?: number;
-  compilation?: boolean;
-  playCount?: number;
-  rating?: number;
-  starred?: boolean;
-  [key: string]: unknown;
-}
-
-export interface RawArtist {
-  id: string;
-  name: string;
-  albumCount: number;
-  songCount: number;
-  genres?: string[];
-  biography?: string;
-  playCount?: number;
-  rating?: number;
-  starred?: boolean;
-  [key: string]: unknown;
-}
-
-
-export interface RawPlaylist {
-  id: string;
-  name: string;
-  comment?: string;
-  public: boolean;
-  songCount: number;
-  duration?: number;
-  owner: string;
-  [key: string]: unknown;
-}
-
-/**
- * Format duration from seconds to MM:SS format
- * @param seconds Duration in seconds
- * @returns Formatted string like "3:45"
- */
-export function formatDuration(seconds?: number): string {
-  if (seconds === null || seconds === undefined || seconds <= 0) {
-    return '0:00';
-  }
-
-  const minutes = Math.floor(seconds / 60);
-  const remainingSeconds = Math.floor(seconds % 60);
-  return `${minutes}:${remainingSeconds.toString().padStart(2, '0')}`;
-}
-
-/**
- * Extract the primary genre from various formats
- * @param rawSong The raw song data
- * @returns The primary genre name or undefined
- */
-function extractGenre(rawSong: RawSong): string | undefined {
-  // Try genres array first (newer format)
-  if (rawSong.genres && Array.isArray(rawSong.genres) && rawSong.genres.length > 0) {
-    const firstGenre = rawSong.genres[0];
-    if (firstGenre) {
-      return firstGenre.name;
-    }
-  }
-  // Fall back to genre string
-  if (rawSong.genre !== null && rawSong.genre !== undefined && rawSong.genre !== '') {
-    return rawSong.genre;
-  }
-  return undefined;
-}
-
-/**
- * Extract all genres from the raw song data
- * @param rawSong The raw song data
- * @returns Array of genre names or undefined
- */
-function extractAllGenres(rawSong: RawSong): string[] | undefined {
-  // Try genres array first (newer format)
-  if (rawSong.genres && Array.isArray(rawSong.genres) && rawSong.genres.length > 0) {
-    return rawSong.genres.map(g => g.name).filter(Boolean);
-  }
-  // Fall back to single genre string as array
-  if (rawSong.genre !== null && rawSong.genre !== undefined && rawSong.genre !== '') {
-    return [rawSong.genre];
-  }
-  return undefined;
 }
 
 
@@ -210,127 +114,3 @@ export function transformSongsToDTO(rawSongs: unknown): SongDTO[] {
   return rawSongs.map((song) => transformToSongDTO(song as RawSong));
 }
 
-export function transformToAlbumDTO(rawAlbum: RawAlbum): AlbumDTO {
-  const dto: AlbumDTO = {
-    id: rawAlbum.id,
-    name: rawAlbum.name || '',
-    artist: rawAlbum.artist || '',
-    artistId: rawAlbum.artistId,
-    songCount: rawAlbum.songCount || 0,
-    durationFormatted: formatDuration(rawAlbum.duration),
-  };
-
-  if (rawAlbum.albumArtist !== undefined) {
-    dto.albumArtist = rawAlbum.albumArtist;
-  }
-
-  if (rawAlbum.albumArtistId !== undefined) {
-    dto.albumArtistId = rawAlbum.albumArtistId;
-  }
-
-  if (rawAlbum.releaseYear !== undefined) {
-    dto.releaseYear = rawAlbum.releaseYear;
-  }
-
-  const genre = extractGenre(rawAlbum as unknown as RawSong);
-  if (genre !== undefined) {
-    dto.genre = genre;
-  }
-
-  const genres = extractAllGenres(rawAlbum as unknown as RawSong);
-  if (genres !== undefined) {
-    dto.genres = genres;
-  }
-
-  if (rawAlbum.compilation !== undefined) {
-    dto.compilation = rawAlbum.compilation;
-  }
-
-  if (rawAlbum.playCount !== undefined) {
-    dto.playCount = rawAlbum.playCount;
-  }
-
-  if (rawAlbum.rating !== undefined) {
-    dto.rating = rawAlbum.rating;
-  }
-
-  if (rawAlbum.starred !== undefined) {
-    dto.starred = rawAlbum.starred;
-  }
-
-  return dto;
-}
-
-export function transformAlbumsToDTO(rawAlbums: unknown): AlbumDTO[] {
-  if (!Array.isArray(rawAlbums)) {
-    return [];
-  }
-
-  return rawAlbums.map((album) => transformToAlbumDTO(album as RawAlbum));
-}
-
-export function transformToArtistDTO(rawArtist: RawArtist): ArtistDTO {
-  const dto: ArtistDTO = {
-    id: rawArtist.id,
-    name: rawArtist.name || '',
-    albumCount: rawArtist.albumCount || 0,
-    songCount: rawArtist.songCount || 0,
-  };
-
-  if (rawArtist.genres !== undefined) {
-    dto.genres = rawArtist.genres;
-  }
-
-  if (rawArtist.biography !== undefined) {
-    dto.biography = rawArtist.biography;
-  }
-
-  if (rawArtist.playCount !== undefined) {
-    dto.playCount = rawArtist.playCount;
-  }
-
-  if (rawArtist.rating !== undefined) {
-    dto.rating = rawArtist.rating;
-  }
-
-  if (rawArtist.starred !== undefined) {
-    dto.starred = rawArtist.starred;
-  }
-
-  return dto;
-}
-
-export function transformArtistsToDTO(rawArtists: unknown): ArtistDTO[] {
-  if (!Array.isArray(rawArtists)) {
-    return [];
-  }
-
-  return rawArtists.map((artist) => transformToArtistDTO(artist as RawArtist));
-}
-
-
-
-export function transformToPlaylistDTO(rawPlaylist: RawPlaylist): PlaylistDTO {
-  const dto: PlaylistDTO = {
-    id: rawPlaylist.id,
-    name: rawPlaylist.name || '',
-    public: rawPlaylist.public || false,
-    songCount: rawPlaylist.songCount || 0,
-    durationFormatted: formatDuration(rawPlaylist.duration),
-    owner: rawPlaylist.owner || '',
-  };
-
-  if (rawPlaylist.comment !== undefined) {
-    dto.comment = rawPlaylist.comment;
-  }
-
-  return dto;
-}
-
-export function transformPlaylistsToDTO(rawPlaylists: unknown): PlaylistDTO[] {
-  if (!Array.isArray(rawPlaylists)) {
-    return [];
-  }
-
-  return rawPlaylists.map((playlist) => transformToPlaylistDTO(playlist as RawPlaylist));
-}
